Update README instructions for Docker Compose
Documented instructions for using Spring Boot Docker
Compose support to run the guide. This is the preferred
option for runnign the guide. Additional instructions for homebrew
installation were removed.

A section was added about running the application as a jar
and as a container. Native compilation was excluded as it would
require a custom hint because of how the guide is written.
Native was excluded as it was not simple to implement and custom
hints are beyond the scope of this guide.