Docker Compose Guide Update (#31)
* Add code in complete required for Docker Compose

This commit adds dependencies and a Docker Compose file
that is required to change the guide to use Docker compose
support. In this commit, only the complete folder is considered.

* Update README instructions for Docker Compose

Documented instructions for using Spring Boot Docker
Compose support to run the guide. This is the preferred
option for runnign the guide. Additional instructions for homebrew
installation were removed.

A section was added about running the application as a jar
and as a container. Native compilation was excluded as it would
require a custom hint because of how the guide is written.
Native was excluded as it was not simple to implement and custom
hints are beyond the scope of this guide.

* Update Initial folder to represent empty project

The initial folder is modified so that it is a true
representation of what will be downloaded from Spring
Initializr. This includes adding the new Docker Compose
dependency as well as a compose.yaml file.

* Remove reference to a web application

* Update typo in README

* Fix grammatical errors in README